requirements.txt
pyrogram
tgcrypto
motor
dnspython
requests
python-dotenv
pymongo
groq
aiohttp
pytz
aiohttp
